requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
huggingface_hub==0.19.4
transformers==4.36.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.4
scikit-learn==1.3.2
requests==2.32.3
feedparser==6.0.10
beautifulsoup4==4.12.2
langdetect==1.0.9
matplotlib==3.7.2
streamlit==1.37.1
